+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Parser
   - **_Breaking Change_**: add `ASTNode.token`.
   - **_Breaking Change_**: `IParserBuilder.useLexer` must be called and only can be called once to set the lexer, instead of setting the lexer in `IParserBuilder.build`.
+  - **_Breaking Change_**: `IParserBuilder.useData` can only be called once.
   - Feat: add `IParserBuilder.useData`.
 - Optimize package size.